+++ src/index.ts
@@ -9,21 +9,33 @@ import { ActivityTypesEx, MessagingExtensionQuery, MessagingExtensionResult } fr
  */
 export interface IMessagingExtensionMiddlewareProcessor {
     /**
-     * Processes incoming queries
+     * Processes incoming queries (composeExtension/query)
      * @param context the turn context
      * @param value the value of the query
      */
-    onQuery(context: TurnContext, value: MessagingExtensionQuery): Promise<MessagingExtensionResult>;
+    onQuery?(context: TurnContext, value: MessagingExtensionQuery): Promise<MessagingExtensionResult>;
     /**
-     * Process incoming requests for Messaging Extension settings
+     * Process incoming requests for Messaging Extension settings (composeExtension/querySettingUrl)
      * @param context the turn context
      */
-    onQuerySettingsUrl(context: TurnContext): Promise<{ title: string, value: string }>;
+    onQuerySettingsUrl?(context: TurnContext): Promise<{ title: string, value: string }>;
     /**
-     * Processes incoming setting updates
+     * Processes incoming setting updates (composeExtension/setting)
      * @param context the turn context
      */
-    onSettings(context: TurnContext): Promise<void>;
+    onSettings?(context: TurnContext): Promise<void>;
+    /**
+     * Processes incoming link queries (composeExtension/queryLink)
+     * @param context the turn context
+     * @param value the value of the query
+     */
+    onQueryLink?(context: TurnContext, value: MessagingExtensionQuery): Promise<MessagingExtensionResult>;
+    /**
+     * Processes incoming link actions (composeExtension/submitAction)
+     * @param context the turn context
+     * @param value the value of the query
+     */
+    onSubmitAction?(context: TurnContext, value: MessagingExtensionQuery): Promise<MessagingExtensionResult>;
 }
 
 /**
@@ -46,7 +58,8 @@ export class MessagingExtensionMiddleware implements Middleware {
     public async onTurn(context: TurnContext, next: () => Promise<void>): Promise<void> {
         switch (context.activity.name) {
             case "composeExtension/query":
-                if (this.commandId === context.activity.value.commandId || this.commandId === undefined) {
+                if ((this.commandId === context.activity.value.commandId || this.commandId === undefined) &&
+                    this.processor.onQuery) {
                     try {
                         const result = await this.processor.onQuery(context, context.activity.value);
                         context.sendActivity({
@@ -67,10 +80,12 @@ export class MessagingExtensionMiddleware implements Middleware {
                             },
                         });
                     }
+                    return;
                 }
                 break;
             case "composeExtension/querySettingUrl":
-                if (this.commandId === context.activity.value.commandId || this.commandId === undefined) {
+                if ((this.commandId === context.activity.value.commandId || this.commandId === undefined) &&
+                    this.processor.onQuerySettingsUrl) {
                     try {
                         const result = await this.processor.onQuerySettingsUrl(context);
                         context.sendActivity({
@@ -99,10 +114,12 @@ export class MessagingExtensionMiddleware implements Middleware {
                             },
                         });
                     }
+                    return;
                 }
                 break;
             case "composeExtension/setting":
-                if (this.commandId === context.activity.value.commandId || this.commandId === undefined) {
+                if ((this.commandId === context.activity.value.commandId || this.commandId === undefined) &&
+                    this.processor.onSettings) {
                     try {
                         await this.processor.onSettings(context);
                         context.sendActivity({
@@ -120,6 +137,59 @@ export class MessagingExtensionMiddleware implements Middleware {
                             },
                         });
                     }
+                    return;
+                }
+                break;
+            case "composeExtension/queryLink":
+                if ((this.commandId === context.activity.value.commandId || this.commandId === undefined) &&
+                    this.processor.onQueryLink) {
+                    try {
+                        const result = await this.processor.onQueryLink(context, context.activity.value);
+                        context.sendActivity({
+                            type: ActivityTypesEx.InvokeResponse,
+                            value: {
+                                body: {
+                                    composeExtension: result,
+                                },
+                                status: 200,
+                            },
+                        });
+                    } catch (err) {
+                        context.sendActivity({
+                            type: ActivityTypesEx.InvokeResponse,
+                            value: {
+                                body: err,
+                                status: 500,
+                            },
+                        });
+                    }
+                    return;
+                }
+                break;
+            case "composeExtension/submitAction":
+                if ((this.commandId === context.activity.value.commandId || this.commandId === undefined) &&
+                    this.processor.onSubmitAction) {
+                    try {
+                        const result = await this.processor.onSubmitAction(context, context.activity.value);
+                        context.sendActivity({
+                            type: ActivityTypesEx.InvokeResponse,
+                            value: {
+                                body: {
+                                    composeExtension: result,
+                                },
+                                status: 200,
+                            },
+                        });
+                    } catch (err) {
+                        context.sendActivity({
+                            type: ActivityTypesEx.InvokeResponse,
+                            value: {
+                                body: err,
+                                status: 500,
+                            },
+                        });
+                    }
+                    return;
                 }
                 break;
             default: