+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [*Unreleased*] - <*2019-03-29*>
+## [*Unreleased*] - <*2019-04-29*>
+
+### Added
+* Added support for Link unfurling (`onQueryLink` - `composeExtension/queryLink`)
+* Added support for Message Actions (`onSubmitAction` - `composeExtension/submitAction`)
+
+### Changes
+* Made all methods of `IMessagingExtensionMiddlewareProcessor` optional
+
+## [*1.0.0*] - <*2019-03-29*>
 
 ### Added
 * Initial release